[#12] postId 타입을 long, BIGINT로 변경

+++ src/main/java/com/atoz/post/PostMapper.java
@@ -20,5 +20,5 @@ public interface PostMapper {
 
     void deletePost(DeletePostRequestDto deletePostRequestDto);
 
-    Optional<OpenPostResponseDto> findPostByPostId(int postId);
+    Optional<OpenPostResponseDto> findPostByPostId(long postId);
 }
